refactor: address fractional task count review

+++ examples/custom_worker_url_routing.rs
@@ -169,9 +169,7 @@ fn cached_file_scan_desired_task_count_handler(
     ev: DesiredTaskCountEvent,
 ) -> Option<Result<DesiredTaskCountEventResponse>> {
     ev.plan.downcast_ref::<CacheExec>()?;
-    Some(Ok(DesiredTaskCountEventResponse::desired(
-        usize::MAX as f64,
-    )))
+    Some(Ok(DesiredTaskCountEventResponse::desired(f64::MAX)))
 }
 
 fn cached_file_scan_scale_up_leaf_node_handler(
